requirements.txt
matplotlib==3.4.2
scikit-learn==0.24.2
numpy==1.21.1
tqdm==4.61.2
opencv-python==4.5.3.56
Pillow==8.3.1